LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/track/track-element/vtt-cue-float-precision-expected.txt


FAIL Float precision of VTTCue attributes line, position and size, stored as floats assert_equals: expected 1.000000000000004 but got 1
